refactor: Split time conversion and display logic into new components and utilities, and clean up project configuration files.
